backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
# openai-whisper==20231117
torch>=2.7.1
torchaudio>=2.7.1
pydub==0.25.1
ffmpeg-python==0.2.0
python-multipart==0.0.6
mysql-connector-python==8.0.33
python-dotenv==1.0.0
